Create logging.py
