+++ sem/__init__.py
@@ -6,6 +6,7 @@ from .gridrunner import BUILD_GRID_PARAMS, SIMULATION_GRID_PARAMS
 from .database import DatabaseManager
 from .utils import list_param_combinations, automatic_parser, stdout_automatic_parser, only_load_some_files, parse_log_components, convert_environment_str_to_dict
 from .cli import cli
+from .logging import process_logs, filter_logs
 
 __all__ = ('CampaignManager', 'SimulationRunner', 'ParallelRunner', 'LptRunner',
            'DatabaseManager', 'list_param_combinations', 'automatic_parser',
